Move common storage tests to one class in src folder
