+++ asto-core/pom.xml
@@ -140,6 +140,16 @@ SOFTWARE.
         </exclusion>
       </exclusions>
     </dependency>
+    <dependency>
+      <groupId>org.junit.jupiter</groupId>
+      <artifactId>junit-jupiter-params</artifactId>
+      <version>5.6.2</version>
+    </dependency>
+    <dependency>
+      <groupId>org.testng</groupId>
+      <artifactId>testng</artifactId>
+      <version>7.3.0</version>
+    </dependency>
     <!-- tests -->
     <!-- s3 mocks deps -->
     <dependency>
